refactored dropdown
