+++ modules/browser_object_panel_ui.py
@@ -6,8 +6,8 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as EC
 
 from modules.page_base import BasePage
-from modules.page_object_about_prefs import AboutPrefs
 from modules.util import BrowserActions, PomUtils
+from modules.components.dropdown import Dropdown
 
 
 class PanelUi(BasePage):
@@ -175,7 +175,7 @@ class PanelUi(BasePage):
 
             with self.driver.context(self.driver.CONTEXT_CONTENT):
                 dropdown_root = self.get_element("clear-history-dropdown")
-                dropdown = AboutPrefs(self.driver).Dropdown(
+                dropdown = Dropdown(
                     page=self, root=dropdown_root, require_shadow=False
                 )
                 dropdown.select_option(option)
