+++ tests/language_packs/test_language_pack_install_addons.py
@@ -2,6 +2,7 @@ import pytest
 from selenium.webdriver import Firefox
 
 from modules.page_object import AboutAddons, AboutPrefs, AmoLanguages
+from modules.components.dropdown import Dropdown
 
 LANGUAGES = [
     (
@@ -56,7 +57,7 @@ def test_language_pack_install_from_addons(
     # perform language changing and assertions in about_prefs
     about_prefs = AboutPrefs(driver, category="general").open()
     language_dropdown = about_prefs.get_element("language-dropdown")
-    dropdown = about_prefs.Dropdown(page=about_prefs, root=language_dropdown)
+    dropdown = Dropdown(page=about_prefs, root=language_dropdown)
     dropdown.select_option(drop_down_name, double_click=True, wait_for_selection=False)
 
     about_prefs.custom_wait(timeout=15).until(
